# feat: add basic in memory caching for module boundary traversal