--- a/packages/babel-plugin/src/index.tsx
+++ b/packages/babel-plugin/src/index.tsx
@@ -2,6 +2,7 @@ import { declare } from '@babel/helper-plugin-utils';
 import template from '@babel/template';
 import * as t from '@babel/types';
 import { importSpecifier } from './utils/ast-builders';
+import { Cache } from './utils/Cache';
 import { visitCssPropPath } from './css-prop';
 import { visitStyledPath } from './styled';
 import { State } from './types';
@@ -13,6 +14,7 @@ export default declare<State>((api) => {
     inherits: require('babel-plugin-syntax-jsx'),
     pre() {
       this.sheets = {};
+      this.cache = new Cache(this.opts);
     },
     visitor: {
       Program: {