--- a/packages/babel-plugin/src/utils/Cache.tsx
+++ b/packages/babel-plugin/src/utils/Cache.tsx
@@ -21,10 +21,24 @@ export class Cache {
     this._cache = new Map();
   }
 
+  /**
+   * Returns hash of `cacheKey` and `namespace` if `namespace` is present
+   * otherwise returns hash of `cacheKey` only.
+   *
+   * @param cacheKey Key for caching
+   * @param namespace Namespace for grouping
+   */
   static getUniqueKey(cacheKey: string, namespace?: string) {
     return hash(namespace ? `${namespace}----${cacheKey}` : cacheKey);
   }
 
+  /**
+   * Lazy evaluates the passed value and save it into the cache.
+   * Also returns the lazy evaluated value.
+   *
+   * @param uniqueKey Unique cache key
+   * @param value Value to be cached
+   */
   _saveToCache<T>(uniqueKey: string, value: () => T) {
     const lazyValue = value();
 
@@ -33,6 +47,10 @@ export class Cache {
     return lazyValue;
   }
 
+  /**
+   * Deletes least recently used value (first in queue) from cache if cache size
+   * reaches its max size.
+   */
   _tryDeletingLRUCachedValue() {
     if (this._cache.size >= this._options.maxSize) {
       const key = this._cache.keys().next().value;
@@ -41,12 +59,24 @@ export class Cache {
     }
   }
 
+  /**
+   * Moves frequently accessed value last into queue so that they won't get deleted
+   * when cache size reaches its max size.
+   *
+   * @param uniqueKey Unique cache key
+   * @param cacheValue Cached value
+   */
   _moveLastInQueue<T>(uniqueKey: string, cacheValue: T) {
     this._cache.delete(uniqueKey);
 
     this._cache.set(uniqueKey, cacheValue);
   }
 
+  /**
+   * Loads cached value from cache and returns it. Also move it last into queue.
+   *
+   * @param uniqueKey Unique cache key
+   */
   _loadFromCache<T>(uniqueKey: string) {
     const cacheValue = this._cache.get(uniqueKey);
 
@@ -55,10 +85,25 @@ export class Cache {
     return cacheValue as T;
   }
 
+  /**
+   * Initialize cache with its options.
+   *
+   * @param options Cache options
+   */
   initialize(options: CacheOptions) {
     this._options = { ...defaultOptions, ...options };
   }
 
+  /**
+   * Checks if we can cache the value. If not, it will just evaluate the passed
+   * value and returns it. If yes, it will start caching it. If value is already
+   * cached, returns it otherwise put it into the cache and delete least recently
+   * used value if cache is full.
+   *
+   * @param cacheKey Key for caching
+   * @param namespace Namespace for grouping
+   * @param value Value to be cached
+   */
   load<T>({
     cacheKey,
     namespace,
@@ -83,14 +128,23 @@ export class Cache {
     return this._saveToCache<T>(uniqueKey, value);
   }
 
+  /**
+   * Returns cache size
+   */
   getSize() {
     return this._cache.size;
   }
 
+  /**
+   * Returns cache keys
+   */
   getKeys() {
     return this._cache.keys();
   }
 
+  /**
+   * Returns cache values
+   */
   getValues() {
     return this._cache.values();
   }