--- a/packages/babel-plugin/src/utils/__mocks__/Cache.tsx
+++ b/packages/babel-plugin/src/utils/__mocks__/Cache.tsx
@@ -1,0 +1,6 @@
+const { Cache } = jest.requireActual('../Cache');
+
+Cache.getUniqueKey = (key: string, namespace?: string) =>
+  namespace ? `${namespace}----${key}` : key;
+
+export { Cache };