--- a/packages/tsconfig.options.json
+++ b/packages/tsconfig.options.json
@@ -19,5 +19,5 @@
     "declaration": true,
     "allowJs": true
   },
-  "exclude": ["**/**.test.tsx", "**/dist"]
+  "exclude": ["**/**.test.tsx", "**/dist", "**/__perf__/**"]
 }