This closes #292 

> @madou: I need early feedback on this one. Creating a draft PR for now. It's still WIP. I might add few more things.

Right now we are saving cache in memory. I left some **TODO's** where we can add file related caching. What I am thinking 🤔, probably we can create a mapping json file with content hash keys and module number as values. These modules will have their own json files with ast info in it.

For eg.

`.cache/compiled/module-mapping.json` will have following content:

```json
{
  "8haj1j": 1,
  "34fgj4": 2
}
```

Each hash value will have separate json file with ast info in it

`.cache/compiled/1.json` will have following content

```json
{
  "type": "File",
  "start": 0,
  "end": 2831,
  "loc": {
    "start": {
      "line": 1,
      "column": 0
    },
    "end": {
      "line": 141,
      "column": 0
    }
  }
...
```
and so on with other modules.

Each key is hash of **module path** and **module code**. Right now hash of **module path** and **module code** will be new when **module code** changes and new entry will be created in `module-mapping.json` file. We can optimize it further with separate hash for both **module path** and **module code** and update its content only when **module code** changes instead of creating new entry in `module-mapping.json`. Didn't give much thought on this one, but this should work.

I did some benchmarking, and here is the result when my laptop was in idle state:

```bash
❯ yarn perf
yarn run v1.22.5
$ ts-node --files ./src/__perf__/module-traversal-caching
#initial-run x 173 ops/sec ±5.23% (72 runs sampled)
#with-cache x 231 ops/sec ±2.28% (76 runs sampled)
#no-cache x 208 ops/sec ±2.41% (77 runs sampled)
Fastest is #with-cache
✨  Done in 24.18s.
```

You can see `#with-cache` is faster than `#no-cache` when we are caching only 2 modules with limited code. Probably it can be much faster than that in real world scenario. I also creating one test suite for `#initial-run` for consistent subsequent results. Even if you swap `#with-cache` position with `#no-cache`, result is same:

```bash
❯ yarn perf
yarn run v1.22.5
$ ts-node --files ./src/__perf__/module-traversal-caching
#initial-run x 173 ops/sec ±4.93% (70 runs sampled)
#no-cache x 202 ops/sec ±1.95% (74 runs sampled)
#with-cache x 234 ops/sec ±2.10% (76 runs sampled)
Fastest is #with-cache
✨  Done in 24.19s.
```