@pgmanutd 2020-09-21
@madou: Damn!! new perf numbers looks amazing
- when I moved cached instance out of `pre()`
- when I started caching module code
- when I started caching module ast traversal

> We will go for file modification time instead of file code. We will append modification time along with module path as `cacheKey`. What you say? 😄

**Previous numbers were:**
```bash
❯ yarn perf
yarn run v1.22.5
$ ts-node --files ./src/__perf__/module-traversal-caching
#initial-run x 173 ops/sec ±5.23% (72 runs sampled)
#with-cache x 231 ops/sec ±2.28% (76 runs sampled)
#no-cache x 208 ops/sec ±2.41% (77 runs sampled)
Fastest is #with-cache
✨  Done in 24.18s.
```

**New numbers are:**
```bash
❯ yarn perf
yarn run v1.22.5
$ ts-node --files ./src/__perf__/module-traversal-caching
#initial-run x 173 ops/sec ±5.19% (70 runs sampled)
#with-cache x 292 ops/sec ±1.85% (79 runs sampled)
#no-cache x 210 ops/sec ±0.93% (78 runs sampled)
Fastest is #with-cache
✨  Done in 24.53s.
```